feat: Add new screen for viewing reports
- Added a new screen `ViewReport` to the `MainNavigator`
- Updated the `MainLayout` component to include a dark status bar
- Updated the `CompartmentInfo` screen to include a
`KeyboardAvoidingView`
- Changed the width of elements in the `CompartmentInfo` screen
- Reordered and styled buttons in the `CompartmentInfo` screen
- Disabled buttons while in "edit" mode
- Updated the app icon to 'oilTruck.png'
- Added new images: `gasStation.png` and `tabIcon.png`
- Updated app version to 1.0.1
- Added dependencies: `date-fns`, `react-native-uuid`, and `@types/uuid`

The new `ViewReport` screen allows users to view reports. This addition
enhances the functionality of the app by providing an easy way to access
and review reports. The changes to the `MainLayout` component ensure a
consistent and better user experience with the addition of a dark status
bar. The `KeyboardAvoidingView` in the `CompartmentInfo` screen helps
avoid text input overlap with the keyboard. The updated styling and
layout in the `CompartmentInfo` screen improve the visual presentation
and usability. The new images and updated app icon contribute to a more
visually appealing interface.